perf(retrieval): optimize score normalization loops using standard comparison operators
Replaced f32::min and f32::max calls with standard `<` and `>` comparison operators in normalize_scores, normalize_scores_in_place, and merge_results. This bypasses IEEE 754 NaN and negative zero propagation branch overhead, allowing LLVM compiler auto-vectorization. Added cargo-mutants ignores for equivalent comparison mutants. Condense tests to satisfy LOC gate while restoring full comments and test named variables.
