+++ crates/csm-cli/Cargo.toml
@@ -13,7 +13,7 @@ csm-core-lib = { path = "../csm-core-lib", version = "0.3.7" }
 csm-memory = { path = "../csm-memory", version = "0.3.7" }
 csm-embedding = { path = "../csm-embedding", version = "0.3.7" }
 csm-retrieval = { path = "../csm-retrieval", version = "0.3.7" }
-csm-persistence = { path = "../csm-persistence", version = "0.3.7", default-features = false }
+csm-persistence = { path = "../csm-persistence", version = "0.3.7" }
 clap = { version = "4.5", features = ["derive", "env", "string"] }
 clap_complete = { version = "4.5.42" }
 anyhow = "1.0.102"
