+++ crates/csm-memory/Cargo.toml
@@ -23,7 +23,7 @@ rand = { workspace = true, optional = true }
 rand_chacha = { version = "0.10.0", optional = true }
 
 [target.'cfg(not(target_arch = "wasm32"))'.dependencies]
-rayon = { version = "1.10", optional = true }
+rayon = "1.10"
 
 [target.'cfg(target_arch = "wasm32")'.dependencies]
 js-sys = "0.3.77"
@@ -33,7 +33,7 @@ tokio = { workspace = true, features = ["macros", "rt-multi-thread"] }
 
 [features]
 default = []
-parallel = ["csm-core-lib/parallel", "dep:rayon"]
+parallel = ["csm-core-lib/parallel"]
 ann-hnsw = ["dep:hnsw_rs", "dep:tempfile", "dep:bincode"]
 ann-lsh = ["dep:bincode", "dep:rand", "dep:rand_chacha"]
