+++ crates/csm-persistence/Cargo.toml
@@ -26,7 +26,7 @@ tokio = { workspace = true, features = ["macros", "rt-multi-thread"] }
 tempfile = { workspace = true }
 
 [features]
-default = []
+default = ["persistence"]
 persistence = ["dep:libsql"]
 libsql = ["persistence"]
 wasm = ["uuid"]
