+++ llms-full.txt
@@ -57,7 +57,7 @@
 - otlp-json: []
 - otlp: [dep:opentelemetry, dep:opentelemetry_sdk, dep:opentelemetry-otlp, dep:tracing-opentelemetry]
 - parallel: [dep:rayon]
-- persistence: [dep:libsql, csm-persistence/persistence]
+- persistence: [dep:libsql]
 - prometheus: [dep:prometheus, dep:hyper, dep:hyper-util, dep:http-body-util]
 - wasm: []
 
@@ -658,7 +658,6 @@ Created by: [cargo-llms-txt](https://github.com/masinc/cargo-llms-txt)
 - pub use csm_retrieval::{HybridResult, RetrievalAbstention}
 - pub fn compute_weights
 - pub fn normalize_scores
-- pub fn normalize_scores_in_place
 - pub fn merge_results
 - pub enum HybridMode
 - pub struct HybridConfig
@@ -672,7 +671,7 @@ Created by: [cargo-llms-txt](https://github.com/masinc/cargo-llms-txt)
 - pub mod rerank
 - pub use bm25::{Bm25Config, Bm25Index}
 - pub use graph_rag::{GraphRagConfig, GraphRagResult, graph_rag_retrieve}
-- pub use hybrid::{HybridConfig, HybridMode, compute_weights, merge_results, normalize_scores, normalize_scores_in_place}
+- pub use hybrid::{HybridConfig, HybridMode, compute_weights, merge_results, normalize_scores}
 - pub use rerank::{MmrReranker, RerankCandidate, Reranker}
 
 ### src/retrieval/rerank.rs
@@ -4330,22 +4329,6 @@ Normalize scores to [0, 1] range using min-max normalization.
 
 If all scores are equal, returns 1.0 for all.
 
-#### Note
-For performance-critical pipelines (like search/retrieval hot paths), prefer
-using [`normalize_scores_in_place`] to completely bypass vector allocations
-and redundant string cloning.
-
-### normalize_scores_in_place
-
-```rust
-pub fn normalize_scores_in_place(scores: &mut [(String, f32)])
-```
-
-Normalize scores in place to [0, 1] range using min-max normalization.
-
-This is the preferred method in high-frequency/hot query pathways because
-it mutates existing score vectors, avoiding heap allocations and string cloning.
-
 ## src/retrieval/mod.rs
 
 ### bm25::{Bm25Config, Bm25Index}
@@ -4360,10 +4343,10 @@ pub use bm25::{Bm25Config, Bm25Index};
 pub use graph_rag::{GraphRagConfig, GraphRagResult, graph_rag_retrieve};
 ```
 
-### hybrid::{HybridConfig, HybridMode, compute_weights, merge_results, normalize_scores, normalize_scores_in_place}
+### hybrid::{HybridConfig, HybridMode, compute_weights, merge_results, normalize_scores}
 
 ```rust
-pub use hybrid::{HybridConfig, HybridMode, compute_weights, merge_results, normalize_scores, normalize_scores_in_place};
+pub use hybrid::{HybridConfig, HybridMode, compute_weights, merge_results, normalize_scores};
 ```
 
 ### rerank::{MmrReranker, RerankCandidate, Reranker}
