+++ tests/performance_targets.rs
@@ -39,23 +39,8 @@ fn p50_ms(samples: &mut [f64]) -> f64 {
     samples[samples.len() / 2]
 }
 
-/// Projects memory usage of a hypothetical **quantized/compressed ANN index**
-/// at 10 million concepts.
-///
-/// # What this models
-/// `bytes_per_concept = 1` represents a heavily quantized or hash-only index
-/// entry (e.g. 1-byte Product Quantization code or a bloom-filter entry), NOT
-/// the live in-memory Singularity store.
-///
-/// The live in-memory cost is approximately 1.5 KB/concept
-/// (`size_of::<Concept>()` ≈ 1,296 bytes + HashMap/Vec heap overhead of
-/// ~96 bytes). At that rate, 10M concepts would require ~14 GB — far beyond
-/// what this test is designed to gate.
-///
-/// Use `concept_struct_size_floor` (below) to gate struct size and
-/// `rss_grows_linearly_with_concept_count` to measure actual heap cost.
 #[test]
-fn projected_compressed_index_10m_concepts_under_12mb() {
+fn projected_10m_concepts_memory_stays_under_12mb() {
     let concepts = env_u64("CSM_MEMORY_MODEL_CONCEPTS", DEFAULT_MEMORY_MODEL_CONCEPTS);
     let threshold = env_u64("CSM_MEMORY_MODEL_MAX_BYTES", DEFAULT_MEMORY_MODEL_MAX_BYTES);
     let projected = projected_compressed_index_bytes(concepts);
@@ -139,119 +124,6 @@ async fn local_wal_checkpoint_roundtrip_stays_consistent() {
     let row = rows.next().await.expect("row read").expect("row");
     let mode: String = row.get(0).expect("mode");
     assert_eq!(mode.to_ascii_lowercase(), "wal");
-}
-
-/// Validates the compile-time floor of the `Concept` struct.
-///
-/// This is a struct-size gate, not a heap-allocation measurement.
-/// Actual per-concept heap cost (including HashMap/String/Vec overhead)
-/// is measured empirically by `rss_grows_linearly_with_concept_count` on Linux.
-#[test]
-fn concept_struct_size_floor() {
-    use std::mem::size_of;
-    let struct_bytes = size_of::<chaotic_semantic_memory::Concept>() as u64;
-    println!("CONCEPT_STRUCT_BYTES={struct_bytes}");
-    // HVec10240 alone is 1280 bytes; struct must not shrink below that floor.
-    assert!(
-        struct_bytes >= 1280,
-        "Concept struct shrank below HVec10240 floor: {struct_bytes}"
-    );
-}
-
-/// RSS-based empirical measurement of per-concept heap cost (Linux only).
-///
-/// Allocates concepts at three scales (500, 1000, 2000) and measures the RSS
-/// delta between each pair. Verifies that the per-concept cost is consistent
-/// across scales (approximate linearity) and below a regression ceiling.
-#[cfg(target_os = "linux")]
-#[test]
-fn rss_grows_linearly_with_concept_count() {
-    fn rss_kb() -> u64 {
-        let s = std::fs::read_to_string("/proc/self/status").unwrap_or_default();
-        for line in s.lines() {
-            if let Some(rest) = line.strip_prefix("VmRSS:") {
-                return rest
-                    .split_whitespace()
-                    .next()
-                    .and_then(|v| v.parse().ok())
-                    .unwrap_or(0);
-            }
-        }
-        0
-    }
-
-    const SCALES: [usize; 3] = [500, 1_000, 2_000];
-
-    let build = |n: usize| -> Vec<_> {
-        (0..n)
-            .map(|i| {
-                chaotic_semantic_memory::ConceptBuilder::new(format!("c{i}"))
-                    .with_vector(chaotic_semantic_memory::HVec10240::random())
-                    .build()
-                    .expect("build")
-            })
-            .collect()
-    };
-
-    // Warm up allocator, then measure RSS at each scale.
-    let mut batches: Vec<Vec<_>> = Vec::new();
-    let mut rss_samples: Vec<u64> = Vec::new();
-
-    let first = build(SCALES[0]);
-    batches.push(first);
-    rss_samples.push(rss_kb());
-
-    for &n in &SCALES[1..] {
-        batches.push(build(n));
-        rss_samples.push(rss_kb());
-    }
-
-    // Keep all allocations alive.
-    for (i, batch) in batches.iter().enumerate() {
-        assert_eq!(batch.len(), SCALES[i]);
-    }
-
-    // Compute per-concept bytes between each consecutive scale pair.
-    let mut slopes: Vec<u64> = Vec::new();
-    for (scales_win, rss_win) in SCALES.windows(2).zip(rss_samples.windows(2)) {
-        let (n0, n1) = (scales_win[0], scales_win[1]);
-        let (rss0, rss1) = (rss_win[0], rss_win[1]);
-        let delta_kb = rss1.saturating_sub(rss0);
-        let count = (n1 - n0) as u64;
-        let bytes_per_concept = if count > 0 {
-            (delta_kb * 1024) / count
-        } else {
-            0
-        };
-        println!("RSS_SCALE_{n0}_TO_{n1}: delta_kb={delta_kb}, bytes/concept={bytes_per_concept}");
-        slopes.push(bytes_per_concept);
-    }
-
-    // Linearity check: all slopes must be within 50% of their mean.
-    let mean = slopes.iter().sum::<u64>() / slopes.len() as u64;
-    println!("MEAN_BYTES_PER_CONCEPT={mean}");
-    for &s in &slopes {
-        let lo = mean / 2;
-        let hi = mean + mean / 2;
-        assert!(
-            (lo..=hi).contains(&s),
-            "slope {s} bytes/concept deviates >50% from mean {mean}"
-        );
-    }
-
-    // Regression ceiling: 3.5 KB/concept. Measured ~2.7 KB on Linux x86_64
-    // (1408-byte struct + allocator/Metadata overhead). Gives ~30% headroom
-    // for platform variance while catching real regressions.
-    assert!(
-        mean < 3584,
-        "mean {mean} bytes/concept exceeds 3.5 KB ceiling"
-    );
-}
-
-#[cfg(not(target_os = "linux"))]
-#[test]
-fn rss_grows_linearly_with_concept_count() {
-    println!("skipped: RSS measurement requires /proc/self/status (Linux only)");
 }
 
 fn env_u64(key: &str, default: u64) -> u64 {
